Correction

+++ 100-clean_web_static.py
@@ -82,7 +82,8 @@ def do_clean(number=0):
     dir = run('ls -1t /data/web_static/releases/').split('\n')
     nb_dir = len(dir)
     while nb_dir > nbr:
-        run('rm -rf /data/web_static/releases/{}/'.format(dir[i]))
+        if dir[i] != 'test':
+            run('rm -rf /data/web_static/releases/{}/'.format(dir[i]))
         i += 1
         nb_dir -= 1
     i = 0
